Make console its own project
Make the console application its own project so that the web app self-contained publish works, as it can't reference another executable.
