+++ src/AdventOfCode.Console/ConsoleLogger.cs
