+++ src/AdventOfCode.Console/Properties/AssemblyInfo.cs
@@ -1,0 +1,6 @@
+﻿// Copyright (c) Martin Costello, 2015. All rights reserved.
+// Licensed under the Apache 2.0 license. See the LICENSE file in the project root for full license information.
+
+using System.Runtime.InteropServices;
+
+[assembly: Guid("3a43dddc-04d3-4691-8e96-f8e5f4b3a553")]
